# More efficient parsing of lists, sets, maps, and unions