Before:
```
[info] Benchmark                            (size)   Mode  Cnt        Score        Error  Units
[info] ListOfBooleansReading.smithy4s          128  thrpt    5   299216.512 ±   2984.556  ops/s
[info] ListOfBooleansWriting.smithy4s          128  thrpt    5  1675363.057 ±  53088.196  ops/s
[info] MapOfIntsToBooleansReading.smithy4s     128  thrpt    5   120887.869 ±   3358.230  ops/s
[info] SetOfIntsReading.smithy4s               128  thrpt    5   146052.002 ±   2970.362  ops/s
```
After:
```
[info] Benchmark                            (size)   Mode  Cnt        Score        Error  Units
[info] ListOfBooleansReading.smithy4s          128  thrpt    5   370941.376 ±   3916.165  ops/s
[info] ListOfBooleansWriting.smithy4s          128  thrpt    5  2617735.800 ± 139036.275  ops/s
[info] MapOfIntsToBooleansReading.smithy4s     128  thrpt    5   160719.243 ±   3958.283  ops/s
[info] SetOfIntsReading.smithy4s               128  thrpt    5   188980.410 ±   2519.980  ops/s
```